+++ aderyn_core/src/ast/impls/disp/events.rs
@@ -1,0 +1,8 @@
+use crate::ast::*;
+use std::fmt::Display;
+
+impl Display for EventDefinition {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!("event {}{}", self.name, self.parameters))
+    }
+}
