+++ aderyn_core/src/ast/impls/disp/identifiers.rs
@@ -1,0 +1,15 @@
+use crate::ast::*;
+use std::fmt::Display;
+use std::hash::{Hash, Hasher};
+
+impl Display for Identifier {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.name.as_str())
+    }
+}
+
+impl Display for IdentifierPath {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.name.as_str())
+    }
+}
