+++ aderyn_core/src/ast/impls/disp/structures.rs
@@ -1,0 +1,16 @@
+use crate::ast::*;
+use crate::visitor::ast_visitor::*;
+use eyre::Result;
+use std::fmt::Display;
+
+impl Display for StructDefinition {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!("struct {} {{\n", self.name))?;
+
+        for member in self.members.iter() {
+            f.write_fmt(format_args!("\t{member};\n"))?;
+        }
+
+        f.write_str("}")
+    }
+}
