+++ aderyn_core/src/ast/impls/disp/types.rs
@@ -1,0 +1,62 @@
+use crate::ast::*;
+
+use std::fmt::Display;
+
+impl Display for TypeName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TypeName::ElementaryTypeName(elementary_type_name) => elementary_type_name.fmt(f),
+            TypeName::UserDefinedTypeName(user_defined_type_name) => user_defined_type_name.fmt(f),
+            TypeName::ArrayTypeName(array_type_name) => array_type_name.fmt(f),
+            TypeName::Mapping(mapping) => mapping.fmt(f),
+            TypeName::Raw(string) => string.fmt(f),
+            _ => unimplemented!(),
+        }
+    }
+}
+
+impl Display for ElementaryTypeName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.name.as_str())?;
+
+        if let Some(state_mutability) = self.state_mutability {
+            if state_mutability != StateMutability::NonPayable {
+                f.write_fmt(format_args!(" {state_mutability}"))?;
+            }
+        }
+
+        Ok(())
+    }
+}
+
+impl Display for UserDefinedTypeName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        if let Some(path_node) = self.path_node.as_ref() {
+            f.write_fmt(format_args!("{path_node}"))
+        } else {
+            f.write_fmt(format_args!("{}", self.name.as_deref().unwrap_or("")))
+        }
+    }
+}
+
+impl Display for ArrayTypeName {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!("{}", self.base_type))?;
+        f.write_str("[")?;
+
+        if let Some(length) = self.length.as_ref() {
+            f.write_fmt(format_args!("{length}"))?;
+        }
+
+        f.write_str("]")
+    }
+}
+
+impl Display for Mapping {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!(
+            "mapping({} => {})",
+            self.key_type, self.value_type
+        ))
+    }
+}
