+++ aderyn_core/src/ast/impls/disp/user_defined_value_types.rs
@@ -1,0 +1,14 @@
+use crate::ast::*;
+use crate::visitor::ast_visitor::ASTConstVisitor;
+use crate::visitor::ast_visitor::Node;
+use eyre::Result;
+use std::fmt::Display;
+
+impl Display for UserDefinedValueTypeDefinition {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!(
+            "type {} is {}",
+            self.name, self.underlying_type,
+        ))
+    }
+}
