+++ aderyn_core/src/ast/impls/disp/using_for_directives.rs
@@ -1,0 +1,18 @@
+use crate::ast::*;
+use crate::visitor::ast_visitor::*;
+use eyre::Result;
+
+use std::fmt::Display;
+
+impl Display for UsingForDirective {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!(
+            "using {:?} for {}",
+            self.library_name,
+            match self.type_name.as_ref() {
+                Some(type_name) => format!("{type_name}"),
+                None => "_".to_string(),
+            }
+        ))
+    }
+}
