+++ aderyn_core/src/ast/impls/node/blocks.rs
@@ -26,15 +26,3 @@ impl Node for Block {
         Ok(())
     }
 }
-
-impl Display for Block {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("{\n")?;
-
-        for statement in self.statements.iter() {
-            f.write_fmt(format_args!("\t{statement};\n"))?;
-        }
-
-        f.write_str("}")
-    }
-}
