+++ aderyn_core/src/ast/impls/node/errors.rs
@@ -21,9 +21,3 @@ impl Node for ErrorDefinition {
         Ok(())
     }
 }
-
-impl Display for ErrorDefinition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("error {}{}", self.name, self.parameters))
-    }
-}
