+++ aderyn_core/src/ast/impls/node/events.rs
@@ -24,9 +24,3 @@ impl Node for EventDefinition {
         Ok(())
     }
 }
-
-impl Display for EventDefinition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("event {}{}", self.name, self.parameters))
-    }
-}
