+++ aderyn_core/src/ast/impls/node/functions.rs
@@ -3,12 +3,6 @@ use crate::visitor::ast_visitor::*;
 use eyre::Result;
 use std::fmt::Display;
 
-impl Display for FunctionKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
-    }
-}
-
 impl Node for ParameterList {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         if visitor.visit_parameter_list(self)? {
@@ -25,22 +19,6 @@ impl Node for ParameterList {
     fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         visitor.visit_node_id(Some(self.id))?;
         Ok(())
-    }
-}
-
-impl Display for ParameterList {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("(")?;
-
-        for (i, parameter) in self.parameters.iter().enumerate() {
-            if i > 0 {
-                f.write_str(", ")?;
-            }
-
-            f.write_fmt(format_args!("{parameter}"))?;
-        }
-
-        f.write_str(")")
     }
 }
 
@@ -72,28 +50,6 @@ impl Node for OverrideSpecifier {
     }
     fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         visitor.visit_node_id(Some(self.id))?;
-        Ok(())
-    }
-}
-
-impl Display for OverrideSpecifier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("override")?;
-
-        if !self.overrides.is_empty() {
-            f.write_str("(")?;
-
-            for (i, identifier_path) in self.overrides.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(", ")?;
-                }
-
-                f.write_fmt(format_args!("{:?}", identifier_path))?;
-            }
-
-            f.write_str(")")?;
-        }
-
         Ok(())
     }
 }
@@ -212,44 +168,5 @@ impl FunctionDefinition {
         }
 
         ids
-    }
-}
-
-impl Display for FunctionDefinition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", self.kind()))?;
-
-        if !self.name.is_empty() {
-            f.write_fmt(format_args!(" {}", self.name))?;
-        }
-
-        f.write_fmt(format_args!("{} {}", self.parameters, self.visibility))?;
-
-        if let Some(state_mutability) = &self.state_mutability {
-            if *state_mutability != StateMutability::NonPayable {
-                f.write_fmt(format_args!(" {}", state_mutability))?;
-            }
-        }
-
-        if let Some(true) = self.r#virtual {
-            f.write_str(" virtual")?;
-        }
-
-        if let Some(overrides) = self.overrides.as_ref() {
-            f.write_fmt(format_args!(" {overrides}"))?;
-        }
-
-        for modifier in self.modifiers.iter() {
-            f.write_fmt(format_args!(" {modifier}"))?;
-        }
-
-        if !self.return_parameters.parameters.is_empty() {
-            f.write_fmt(format_args!(" returns {}", self.return_parameters))?;
-        }
-
-        match self.body.as_ref() {
-            Some(body) => f.write_fmt(format_args!(" {body}")),
-            None => f.write_str(";"),
-        }
     }
 }
