+++ aderyn_core/src/ast/impls/node/modifiers.rs
@@ -31,32 +31,6 @@ impl Node for ModifierDefinition {
     }
 }
 
-impl Display for ModifierDefinition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str("modifier")?;
-
-        if !self.name.is_empty() {
-            f.write_fmt(format_args!(" {}", self.name))?;
-        }
-
-        f.write_fmt(format_args!("{}", self.parameters))?;
-
-        if self.visibility != Visibility::Internal {
-            f.write_fmt(format_args!("{} {}", self.parameters, self.visibility))?;
-        }
-
-        if let Some(true) = self.r#virtual {
-            f.write_fmt(format_args!(" virtual"))?;
-        }
-
-        if let Some(overrides) = self.overrides.as_ref() {
-            f.write_fmt(format_args!(" {overrides}"))?;
-        }
-
-        f.write_fmt(format_args!(" {}", self.body))
-    }
-}
-
 impl IdentifierOrIdentifierPath {
     pub fn get_node_id(&self) -> NodeID {
         match self {
@@ -115,28 +89,6 @@ impl Node for ModifierInvocation {
     }
     fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         visitor.visit_node_id(Some(self.id))?;
-        Ok(())
-    }
-}
-
-impl Display for ModifierInvocation {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{:?}", self.modifier_name))?;
-
-        if let Some(arguments) = self.arguments.as_ref() {
-            f.write_str("(")?;
-
-            for (i, argument) in arguments.iter().enumerate() {
-                if i > 0 {
-                    f.write_str(", ")?;
-                }
-
-                f.write_fmt(format_args!("{argument}"))?;
-            }
-
-            f.write_str(")")?;
-        }
-
         Ok(())
     }
 }
