+++ aderyn_core/src/ast/impls/node/structures.rs
@@ -21,15 +21,3 @@ impl Node for StructDefinition {
         Ok(())
     }
 }
-
-impl Display for StructDefinition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("struct {} {{\n", self.name))?;
-
-        for member in self.members.iter() {
-            f.write_fmt(format_args!("\t{member};\n"))?;
-        }
-
-        f.write_str("}")
-    }
-}
