+++ aderyn_core/src/ast/impls/node/types.rs
@@ -21,19 +21,6 @@ impl Node for TypeName {
     }
 }
 
-impl Display for TypeName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TypeName::ElementaryTypeName(elementary_type_name) => elementary_type_name.fmt(f),
-            TypeName::UserDefinedTypeName(user_defined_type_name) => user_defined_type_name.fmt(f),
-            TypeName::ArrayTypeName(array_type_name) => array_type_name.fmt(f),
-            TypeName::Mapping(mapping) => mapping.fmt(f),
-            TypeName::Raw(string) => string.fmt(f),
-            _ => unimplemented!(),
-        }
-    }
-}
-
 impl Node for ElementaryTypeName {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         visitor.visit_elementary_type_name(self)?;
@@ -53,20 +40,6 @@ impl Hash for ElementaryTypeName {
         self.state_mutability.hash(state);
         self.name.hash(state);
         self.type_descriptions.hash(state);
-    }
-}
-
-impl Display for ElementaryTypeName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.name.as_str())?;
-
-        if let Some(state_mutability) = self.state_mutability {
-            if state_mutability != StateMutability::NonPayable {
-                f.write_fmt(format_args!(" {state_mutability}"))?;
-            }
-        }
-
-        Ok(())
     }
 }
 
@@ -95,16 +68,6 @@ impl Hash for UserDefinedTypeName {
     }
 }
 
-impl Display for UserDefinedTypeName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if let Some(path_node) = self.path_node.as_ref() {
-            f.write_fmt(format_args!("{path_node}"))
-        } else {
-            f.write_fmt(format_args!("{}", self.name.as_deref().unwrap_or("")))
-        }
-    }
-}
-
 impl Node for FunctionTypeName {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         if visitor.visit_function_type_name(self)? {
@@ -127,19 +90,6 @@ impl Node for ArrayTypeName {
     }
 }
 
-impl Display for ArrayTypeName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", self.base_type))?;
-        f.write_str("[")?;
-
-        if let Some(length) = self.length.as_ref() {
-            f.write_fmt(format_args!("{length}"))?;
-        }
-
-        f.write_str("]")
-    }
-}
-
 impl Node for Mapping {
     fn accept(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         if visitor.visit_mapping(self)? {
@@ -147,14 +97,5 @@ impl Node for Mapping {
             self.value_type.accept(visitor)?;
         }
         visitor.end_visit_mapping(self)
-    }
-}
-
-impl Display for Mapping {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!(
-            "mapping({} => {})",
-            self.key_type, self.value_type
-        ))
     }
 }
