+++ aderyn_core/src/ast/impls/node/using_for_directives.rs
@@ -55,16 +55,3 @@ impl Node for UsingForDirective {
         Ok(())
     }
 }
-
-impl Display for UsingForDirective {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!(
-            "using {:?} for {}",
-            self.library_name,
-            match self.type_name.as_ref() {
-                Some(type_name) => format!("{type_name}"),
-                None => "_".to_string(),
-            }
-        ))
-    }
-}
