+++ aderyn_core/src/ast/impls/node/variables.rs
@@ -1,31 +1,6 @@
 use crate::ast::*;
 use crate::visitor::ast_visitor::*;
 use eyre::Result;
-use std::fmt::Display;
-
-impl Display for Mutability {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
-    }
-}
-
-impl Display for StateMutability {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
-    }
-}
-
-impl Display for Visibility {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
-    }
-}
-
-impl Display for StorageLocation {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", format!("{self:?}").to_lowercase()))
-    }
-}
 
 impl VariableDeclaration {
     /// Returns the mutability of the variable that was declared.
@@ -73,40 +48,6 @@ impl Node for VariableDeclaration {
     }
     fn accept_id(&self, visitor: &mut impl ASTConstVisitor) -> Result<()> {
         visitor.visit_node_id(Some(self.id))?;
-        Ok(())
-    }
-}
-
-impl Display for VariableDeclaration {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_fmt(format_args!("{}", self.type_name.as_ref().unwrap()))?;
-
-        if self.storage_location != StorageLocation::Default {
-            f.write_fmt(format_args!(" {}", self.storage_location))?;
-        }
-
-        if let Some(mutability) = self.mutability.as_ref() {
-            if mutability != &Mutability::Mutable {
-                f.write_fmt(format_args!(" {mutability}"))?;
-            }
-        }
-
-        if let Some(true) = self.indexed {
-            f.write_str(" indexed")?;
-        }
-
-        if self.state_variable {
-            f.write_fmt(format_args!(" {}", self.visibility))?;
-        }
-
-        if !self.name.is_empty() {
-            f.write_fmt(format_args!(" {}", self.name))?;
-        }
-
-        if let Some(value) = self.value.as_ref() {
-            f.write_fmt(format_args!(" = {value}"))?;
-        }
-
         Ok(())
     }
 }
